move: directory by layer
