+++ planetary-orbital-calculator/src/main/java/com/tdd/application/TrueAnomalyCalculator.java
@@ -1,4 +1,4 @@
-package com.tdd;
+package com.tdd.application;
 
 public class TrueAnomalyCalculator {
