+++ planetary-orbital-calculator/src/main/java/com/tdd/domain/PlanetOrbit.java
@@ -1,4 +1,4 @@
-package com.tdd;
+package com.tdd.domain;
 
 public class PlanetOrbit extends Orbit {
