+++ planetary-orbital-calculator/src/main/java/com/tdd/util/Clock.java
@@ -1,4 +1,4 @@
-package com.tdd.utils;
+package com.tdd.util;
 
 import java.time.LocalDateTime;
