+++ planetary-orbital-calculator/src/test/java/com/tdd/application/ArgumentOfPeriapsisCalculatorTest.java
@@ -1,4 +1,4 @@
-package com.tdd;
+package com.tdd.application;
 
 import org.junit.jupiter.api.Test;
