+++ planetary-orbital-calculator/src/test/java/com/tdd/application/CurrentOrbitCalculatorTest.java
@@ -1,7 +1,10 @@
-package com.tdd;
+package com.tdd.application;
 
+import com.tdd.application.CurrentOrbitCalculator;
+import com.tdd.domain.Orbit;
+import com.tdd.domain.PlanetOrbit;
 import com.tdd.planet.EARTH;
-import com.tdd.utils.TimeFreezer;
+import com.tdd.util.TimeFreezer;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
