+++ planetary-orbital-calculator/src/test/java/com/tdd/domain/OrbitTest.java
@@ -1,4 +1,4 @@
-package com.tdd;
+package com.tdd.domain;
 
 import org.junit.jupiter.api.Test;
